+++ src/main/java/com/CuteNekoDragon/Core/common/datagen/DataGen.java
@@ -64,6 +64,16 @@ public class DataGen {
         REGISTRATE.addRawLang("tooltip.svo_core.neptunite", "A jet-black crystal that is unusually reflective.");
         REGISTRATE.addRawLang("tooltip.svo_core.lemon_stone", "Some claim the powdered crystal is a dwarvish delicacy.");
         REGISTRATE.addRawLang("tooltip.svo_core.nekoite", "The delicate shards form a tiny pink meadow.");
+        REGISTRATE.addRawLang("tooltip.svo_core.orpiment", "Despite its high toxicity, this Mineral is widely used in manufacturing and folk medicine.");
+        REGISTRATE.addRawLang("tooltip.svo_core.petrified_slime", "This little guy may be 100,000 years old.");
+        REGISTRATE.addRawLang("tooltip.svo_core.thunder_egg", "According to legend, angry thunder spirits would throw these stones at one another.");
+        REGISTRATE.addRawLang("tooltip.svo_core.pyrite", "Commonly known as \"Fool's Gold\".");
+        REGISTRATE.addRawLang("tooltip.svo_core.ocean_stone", "An old legend claims these stones are the mosaics of ancient mermaids.");
+        REGISTRATE.addRawLang("tooltip.svo_core.ghost_crystal", "There is an aura of coldness around this crystal.");
+        REGISTRATE.addRawLang("tooltip.svo_core.jasper", "When polished, this stone becomes attractively luminous. Prized by ancient peoples for thousands of years.");
+        REGISTRATE.addRawLang("tooltip.svo_core.celestine", "Some early life forms had bones made from this.");
+        REGISTRATE.addRawLang("tooltip.svo_core.marble", "A very popular material for sculptures and construction.");
+        REGISTRATE.addRawLang("tooltip.svo_core.sandstone", "A common type of stone with red and brown striations.");
 
         REGISTRATE.addRawLang("tooltip.svo_core.coins", "%s");
         REGISTRATE.addRawLang("tooltip.svo_core.mineral_product", "Mineral Product");